+++ src/libAACdec/rvlc.h
@@ -122,14 +122,14 @@ amm-info@iis.fraunhofer.de
                 bits decoded as there are available */
 #define RVLC_ERROR_RVL_SUM_BIT_COUNTER_BELOW_ZERO_FWD                     \
   0x40000000 /*  RVL-Dec  negative sum-bitcounter during RVL-fwd-decoding \
-                (long+shrt) */
+                (int32_t+shrt) */
 #define RVLC_ERROR_RVL_SUM_BIT_COUNTER_BELOW_ZERO_BWD                     \
   0x20000000 /*  RVL-Dec  negative sum-bitcounter during RVL-fwd-decoding \
-                (long+shrt) */
+                (int32_t+shrt) */
 #define RVLC_ERROR_FORBIDDEN_CW_DETECTED_FWD \
-  0x08000000 /*  RVL-Dec  forbidden codeword detected fwd (long+shrt) */
+  0x08000000 /*  RVL-Dec  forbidden codeword detected fwd (int32_t+shrt) */
 #define RVLC_ERROR_FORBIDDEN_CW_DETECTED_BWD \
-  0x04000000 /*  RVL-Dec  forbidden codeword detected bwd (long+shrt) */
+  0x04000000 /*  RVL-Dec  forbidden codeword detected bwd (int32_t+shrt) */
 
 void CRvlc_Read(CAacDecoderChannelInfo *pAacDecoderChannelInfo,
                 HANDLE_FDK_BITSTREAM bs);
